+++ openapi_schema_pydantic/v3/v3_0_3/discriminator.py
@@ -1,6 +1,6 @@
 from typing import Dict, Optional
 
-from pydantic import BaseModel
+from pydantic import BaseModel, Extra
 
 
 class Discriminator(BaseModel):
@@ -25,6 +25,7 @@ class Discriminator(BaseModel):
     """
 
     class Config:
+        extra = Extra.forbid
         schema_extra = {
             "examples": [
                 {