+++ openapi_schema_pydantic/v3/v3_0_3/encoding.py
@@ -1,6 +1,6 @@
 from typing import Dict, Optional, Union
 
-from pydantic import BaseModel
+from pydantic import BaseModel, Extra
 
 from .reference import Reference
 
@@ -60,6 +60,7 @@ class Encoding(BaseModel):
     """
 
     class Config:
+        extra = Extra.forbid
         schema_extra = {
             "examples": [
                 {