+++ openapi_schema_pydantic/v3/v3_0_3/example.py
@@ -1,6 +1,6 @@
 from typing import Any, Optional
 
-from pydantic import BaseModel
+from pydantic import BaseModel, Extra
 
 
 class Example(BaseModel):
@@ -33,6 +33,7 @@ class Example(BaseModel):
     """
 
     class Config:
+        extra = Extra.forbid
         schema_extra = {
             "examples": [
                 {"summary": "A foo example", "value": {"foo": "bar"}},