+++ openapi_schema_pydantic/v3/v3_0_3/response.py
@@ -1,6 +1,6 @@
 from typing import Dict, Optional, Union
 
-from pydantic import BaseModel
+from pydantic import BaseModel, Extra
 
 from .header import Header
 from .link import Link
@@ -44,6 +44,7 @@ class Response(BaseModel):
     """
 
     class Config:
+        extra = Extra.forbid
         schema_extra = {
             "examples": [
                 {